requirements.txt

Django==3.1.1
django-cors-headers==3.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
sqlparse==0.3.0
